Done mocha testing before Backend code review

+++ Functions/HomePageController/HomePageController.js
@@ -3,22 +3,35 @@ const { Task } = require("../Models/TaskModel.js");
 
 const addTask = async (req, res) => {
   try {
-    const task = await Task.create({
-      ...req.body,
+    const task = req.body;
+    if (!task) {
+      throw new Error("Task cannot be empty");
+    }
+    if (!task.entry) {
+      throw new Error("Task cannot be empty");
+    }
+    const createTask = await Task.create({
+      ...task,
       taskId: new mongoose.Types.ObjectId(),
     });
     res
       .status(201)
-      .json({ statusCode: 201, message: "Task Added!", data: task })
+      .json({ statusCode: 201, message: "Task Added!", data: createTask })
       .end();
   } catch (error) {
-    res.status(200).send({ statusCode: 400, message: error.message });
+    res.status(400).send({ statusCode: 400, message: error.message, data: {} });
   }
 };
 
 const editTask = async (req, res) => {
   try {
     const { taskId, entry } = req.body;
+    if (!taskId) {
+      throw new Error("Task not found!");
+    }
+    if (!entry) {
+      throw new Error("Task cannot be empty");
+    }
     // eslint-disable-next-line
     const updatedTask = await Task.findOneAndUpdate(
       { taskId: taskId },
@@ -27,18 +40,24 @@ const editTask = async (req, res) => {
     );
     res.status(200).send({ statusCode: 200, message: "Task edited" });
   } catch (error) {
-    res.status(200).send({ statusCode: 400, message: error.message });
+    res.status(400).send({ statusCode: 400, message: error.message });
   }
 };
 
 const deleteTask = async (req, res) => {
   try {
     const taskId = req.params.id;
+    if (!taskId) {
+      throw new Error("Task not found!");
+    }
+    const findTaskByTaskId = await Task.findOne({ taskId: taskId });
+    if (!findTaskByTaskId) {
+      throw new Error("Task not found");
+    }
     await Task.findOneAndDelete({ taskId: taskId });
-    res.status(204).send({ statusCode: 204, message: "Task Deleted" });
+    res.status(200).send({ statusCode: 200, message: "Task Deleted" }).end();
   } catch (error) {
-    console.log("Error: ", error);
-    res.status(200).send({ statusCode: 400, message: error.message });
+    res.status(400).send({ statusCode: 400, message: error.message }).end();
   }
 };
 
@@ -53,7 +72,7 @@ const getUserTask = async (req, res) => {
       .status(200)
       .send({ statusCode: 200, message: "User Tasks", data: userTasks });
   } catch (error) {
-    res.status(200).send({ statusCode: 400, message: error.message });
+    res.status(400).send({ statusCode: 400, message: error.message });
   }
 };
 
@@ -64,7 +83,7 @@ const clearUserTasks = async (req, res) => {
     const clearTasks = await Task.findByIdAndDelete(userId);
     res.status(204).send({ statusCode: 204, message: "All Cleared" });
   } catch (error) {
-    res.status(200).send({ statusCode: 400, message: error.message });
+    res.status(400).send({ statusCode: 400, message: error.message });
   }
 };
